+++ src/main/java/roomescape/controller/ReservationController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.dto.ReservationRequestDto;
-import roomescape.domain.entity.Reservation;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequestDto;
 import roomescape.service.ReservationService;
 
 @RestController
