+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.dto.ReservationTimeRequestDto;
-import roomescape.domain.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.service.ReservationTimeService;
 
 @RestController
