+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -1,8 +1,8 @@
 package roomescape.dao;
 
 import java.util.List;
-import roomescape.domain.dto.ReservationTimeRequestDto;
-import roomescape.domain.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeRequestDto;
 
 public interface ReservationTimeDao {
     ReservationTime create(ReservationTimeRequestDto requestDto);
