+++ src/main/java/roomescape/dto/ReservationRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.domain.dto;
+package roomescape.dto;
 
 public record ReservationRequestDto(String name, String date, Long timeId) {
 }
